pyproject.toml
[tool.poetry]
name = "searchright-technical-assignment"
version = "0.1.0"
description = ""
authors = ["Nuung <qlgks1@naver.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
psycopg2-binary = "^2.9.10"
openai = "^1.75.0"
fastapi = "^0.115.12"
uvicorn = {extras = ["standard"], version = "^0.34.2"}
pgvector = "^0.4.1"
psycopg = {extras = ["binary"], version = "^3.2.8"}
python-dotenv = "^1.1.0"
sqlalchemy = "^2.0.40"
pytest = "^8.3.5"
pytest-asyncio = "^0.26.0"
httpx = "^0.28.1"
tiktoken = "^0.9.0"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.5"
pytest-asyncio = "^0.26.0"
black = "^25.1.0"
isort = "^6.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
